requirements.txt
appdirs
asgiref
CacheControl
certifi
chardet
colorama
contextlib2
coreapi
coreschema
distlib
distro
Django
django-cors-headers
django-rest-framework
django-simple-history
djangorestframework
djangorestframework-simplejwt
drf-yasg
et-xmlfile
Faker
html5lib
idna
inflection
ipaddr
itypes
Jinja2
lockfile
MarkupSafe
msgpack
openapi-codec
openpyxl
packaging
pep517
Pillow
progress
psycopg2-binary
PyJWT
pyparsing
python-dateutil
pytoml
pytz
requests
retrying
ruamel.yaml
ruamel.yaml.clib
simplejson
six
sqlparse
toml
ua-parser
uritemplate
urllib3
webencodings
mysqlclient
python-dotenv